Former-commit-id: a96df6d6ad1bfe7b5392f973fcf64173e5ea36ad [formerly 0c11f99805b786370234f05cbd3da6c3efd69337] [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f] [formerly a96df6d6ad1bfe7b5392f973fcf64173e5ea36ad [formerly 0c11f99805b786370234f05cbd3da6c3efd69337] [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f] [formerly adfeb1306751695abe2eaaf23d62ed4f3baa2430 [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f [formerly 56dc338129ed46612df4db1926e0b7e91551b60d]]]]
Former-commit-id: adfeb1306751695abe2eaaf23d62ed4f3baa2430
Former-commit-id: 0371accc473390fe9993728c95b2fbbffe0da7c7 [formerly 60f9feed3efc51b1179e84a664cd1775442ffe07] [formerly f6e448879003172c3aafecc16e87e7783e176970 [formerly 12c854e7cc66f4fd55c067ee414a3797e0cac5d2]]
Former-commit-id: f350b0f60c1e207b1fc2e10edfcd604fb4c9e1cb [formerly baf0eb6291598aa571c8c038963cc4d2e36770bf]
Former-commit-id: cb21a22731e845b7d20edbc065943a7797bb34b0

+++ edexOsgi/com.raytheon.uf.edex.plugin.bufrobs/src/com/raytheon/uf/edex/plugin/bufrobs/AbstractBufrSfcObsDecoder.java
@@ -91,6 +91,7 @@ import com.raytheon.uf.edex.plugin.bufrobs.category.CategoryParser;
  * Apr 21, 2014 2906       bclement     populated wmo header in record for consistency
  * Jun 12, 2014 3229       bclement     default implementation for getTranslationFile() and createStationId()
  *                                      moved processGeneralFields() and processPrecip() from synoptic land decoder
+ * Jul 23, 2014 3410       bclement     location changed to floats
  * 
  * 
  * </pre>
@@ -534,13 +535,13 @@ public abstract class AbstractBufrSfcObsDecoder {
         } else if (baseName.equalsIgnoreCase(SfcObsPointDataTransform.LATITUDE)) {
             Double lat = (Double) getFieldValue(parser, false);
             if (lat != null) {
-                location.assignLatitude(lat);
+                location.assignLatitude(lat.floatValue());
             }
         } else if (baseName
                 .equalsIgnoreCase(SfcObsPointDataTransform.LONGITUDE)) {
             Double lon = (Double) getFieldValue(parser, false);
             if (lon != null) {
-                location.assignLongitude(lon);
+                location.assignLongitude(lon.floatValue());
             }
         } else if (baseName
                 .equalsIgnoreCase(SfcObsPointDataTransform.ELEVATION)) {
